langchain/Dojo_W10/pyproject.toml
[tool.poetry]
name = "dojo-w10"
version = "0.1.0"
description = ""
authors = ["Eazenli <yizhenli233@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.45.1"
python-dotenv = "^1.1.0"
langchain = "^0.3.25"
pypdf = "^5.6.0"
langchain-openai = "^0.3.22"
langchain-community = "^0.3.25"
langchain-mistralai = "^0.2.10"
langchain-chroma = "^0.2.4"
protobuf = "^3.20.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
